project/requirements.txt
python-telegram-bot==13.15
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
google-api-python-client==2.70.0
python-dotenv==1.0.0